intercomfaceid/requirements.txt
requests
paho-mqtt
opencv-python
face-recognition
paho-mqtt
numpy